requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
cmake==3.26.3
filelock==3.12.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.4
Jinja2==3.1.2
lit==16.0.5.post0
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
packaging==23.1
Pillow==9.5.0
PyYAML==6.0
requests==2.31.0
scipy==1.10.1
submitit==1.4.5
sympy==1.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
triton==2.0.0
typing_extensions==4.6.3
urllib3==2.0.2
